+++ package.json
@@ -28,7 +28,6 @@
     "jsonwebtoken": "^8.5.1",
     "morgan": "^1.10.0",
     "multer": "^1.4.5-lts.1",
-    "mysql": "^2.18.1",
     "mysql2": "^2.3.3",
     "typeorm": "^0.3.10",
     "util": "^0.12.4"
