--- src/middlewares/catchAsync.js
@@ -1,12 +1,0 @@
-const catchAsync = (asyncFn) => {
-    return async (req, res, next) => {
-      try {
-        await asyncFn(req, res, next);
-      } 
-      catch (err) {
-        next(err);
-      };
-    }
-  };
-  
-  module.exports = { catchAsync };
