+++ src/middlewares/index.js
@@ -1,9 +1,7 @@
 const { uploadFiles } = require("./uploadFiles");
 const { validators } = require("./validators");
-const { catchAsync } = require("./catchAsync");
 
 module.exports = {
   uploadFiles,
   validators,
-  catchAsync,
 };
