+++ src/models/dataSource.js
@@ -1,4 +1,4 @@
-const { DataSource } = require("typeORM");
+const { DataSource } = require("typeorm");
 
 const database = new DataSource({
   type: process.env.TYPEORM_TYPE,
@@ -6,9 +6,9 @@ const database = new DataSource({
   port: process.env.TYPEORM_PORT,
   username: process.env.TYPEORM_USERNAME,
   password: process.env.TYPEORM_PASSWORD,
-  database: process.env.TYPEORM_DATABASE,
+  database: process.env.TYPEORM_DATABASE
 });
 
-module.exports = {
-  database,
-};
+module.exports = 
+  database
+// };
