+++ src/models/index.js
@@ -1,5 +1,7 @@
-const userDao = require("./userDao");
+const  userDao  = require("./userDao");
+const  database  = require("./dataSource");
 
 module.exports = {
   userDao,
+  database,
 };
