+++ src/services/index.js
@@ -1,5 +1,5 @@
-const userService = require("./userService");
+const  userService  = require("./userService");
 
 module.exports = {
-  userService,
+  userService
 };
