+++ users.csv
@@ -1,4 +1,4 @@
-id,userId,nickname,password,prifileImg,gender,age,created_at
+id,user_id,nickname,password,prifile_image,gender,age,created_at
 1,Rebekah,Johnson,dnddow@14,https://w.namu.la/s/047dddadae835cf34effb34f3088d4a6a958ad498e124c9e16fcd985db20f0639a7a5eef3dcb0373716885363f145a56f7592303d62c6d55bc9f8fd4b33cb84c7753abc45759cffdd24165828b720fef49f4be1bd09f50a051ea6d99b52f671f88edddb543c2d526bef4a8e07535a748,male,40,2022-09-20-14:00
 2,Fabian,Predovic,wecode$37,https://w.namu.la/s/047dddadae835cf34effb34f3088d4a6a958ad498e124c9e16fcd985db20f0639a7a5eef3dcb0373716885363f145a56f7592303d62c6d55bc9f8fd4b33cb84c7753abc45759cffdd24165828b720fef49f4be1bd09f50a051ea6d99b52f671f88edddb543c2d526bef4a8e07535a748,female,30,2022-09-20-14:00
 3,Elenor,Gottlieb,samsung99*&,https://w.namu.la/s/047dddadae835cf34effb34f3088d4a6a958ad498e124c9e16fcd985db20f0639a7a5eef3dcb0373716885363f145a56f7592303d62c6d55bc9f8fd4b33cb84c7753abc45759cffdd24165828b720fef49f4be1bd09f50a051ea6d99b52f671f88edddb543c2d526bef4a8e07535a748,male,20,2022-09-20-14:00
